+++ CHANGELOG.md
@@ -1,9 +1,16 @@
 # Changelog
 
+### 1.0.3
+
+- Renamed some options.
+- More max attempts and timeout to be job specific (default can be set as well).
+- Fix drain when no active queues.
+- Add deleteJobs option to allow keeping jobs data in Redis.
+
 ### 1.0.2
 
-- Move ioredis to peer dep
+- Move ioredis to peer dep.
 
 ### 0.0.0
 
-- Starting out! 
+- Starting out!